+++ v8/protocol/src/org/jetbrains/v8/protocol/input/Handle.java
@@ -1,7 +1,7 @@
 package org.jetbrains.v8.protocol.input;
 
 public interface Handle {
-  long handle();
+  int handle();
 
   String type();
 }
